Working on checkout flow

+++ lib/screens/splash_screen/main_screen/my_cart_screen/checkout_screen/checkout_screen.dart
@@ -8,7 +8,7 @@ import '../../../../../ui_component/app_textfield.dart';
 import '../../../../../utils/nav/nav.dart';
 
 class CheckoutScreen extends StatefulWidget {
-  final int totalAmount;
+  final double totalAmount;
   const CheckoutScreen({super.key, required this.totalAmount});
 
   @override
